refactor: chat components code
